Quantai_Backend/oldbackend/QuantAi-Backend-master/code/requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
aniso8601==7.0.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.7.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.3.0
audeer==2.2.0
audiofile==1.5.0
audmath==1.4.1
autobahn==24.4.2
Automat==24.8.1
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
boto3==1.37.26
botocore==1.37.26
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.3.2
chex==0.1.88
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
crispy==0.8.0
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
dataclasses==0.6
dataclasses-json==0.6.7
dateparser==1.2.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
distro==1.9.0
Django==4.2.17
django-3-jet==1.0.8
django-allauth==65.5.0
django-auditlog==3.2.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-env==0.2.1
django-environ==0.11.2
django-geoip2-extras==4.1
django-import-export==4.3.9
django-js-asset==3.0.1
django-model-utils==5.0.0
django-oauth-toolkit==3.0.1
django-phone-field==1.8.1
django-picklefield==3.3
django-q==1.3.9
django-referrer-policy==1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djhtml==3.0.6
djongo==1.2.31
dnspython==2.7.0
docopt==0.6.2
drf-social-oauth2==3.1.0
editor==1.6.6
elevenlabs==1.50.3
environ==1.0
et_xmlfile==2.0.0
etils==1.11.0
exceptiongroup==1.2.2
executing==2.2.0
fabio==2024.9.0
facebook-sdk==3.1.0
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
flax==0.10.2
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
geoip2==3.0.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
hdf5plugin==5.0.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
import-export==0.3.1
importlib_resources==6.4.5
incremental==24.7.2
inflection==0.5.1
ipinfo==4.2.1
ipinfo-django==2.0.0
ipython==8.12.3
itypes==1.2.0
jax==0.4.37
jaxlib==0.4.36
jedi==0.19.2
jet-bridge-base==1.12.1
Jinja2==3.1.4
jinxed==1.3.0
jiter==0.8.2
jmespath==1.0.1
jsonfield==3.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
jwt==1.4.0
keras==3.7.0
kiwisolver==1.4.7
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.4
libclang==18.1.1
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
maxminddb==2.8.2
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
namex==0.0.8
narwhals==1.19.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onesignal_python_api==2.2.1
openai==1.58.1
openai-whisper==20240930
openapi-codec==1.3.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.10.2
orjson==3.10.12
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
proglog==0.1.10
promise==2.3
prompt-toolkit==2.0.9
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pymemcache==4.0.0
pymongo==4.1.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.1.4
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytube==15.0.0
pytz==2023.3
pywin32==309
PyYAML==6.0.2
pyzmq==26.2.1
redis==3.5.3
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
runs==1.2.2
Rx==1.6.3
s3transfer==0.11.4
safetensors==0.4.5
scipy==1.14.1
service-identity==24.2.0
silx==2.1.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==1.4.54
sqlparse==0.5.3
srt==3.5.3
srt_equalizer==0.1.10
stack-data==0.6.3
streamlit==1.41.1
style==1.1.0
sympy==1.13.1
tablib==3.8.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
tensorstore==0.1.71
termcolor==2.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20250809
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
ua-parser==0.16.1
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
uuid==1.30
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
whitenoise==6.9.0
win32_setctime==1.2.0
wrapt==1.15.0
xlwt==1.3.0
xmod==1.8.1
xraydb==4.5.4
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
